Move text optional component
refs #67
